+++ Source/Utility/Logger.hpp
@@ -12,7 +12,7 @@ public:
 
 	static void LogLine(std::string string) { std::cout << string << std::endl; }
 
-	static void LogLine(int number)
+	static void LogLine(double number)
 	{
 		std::ostringstream stream;
 		stream << number;
